requirements.txt
django_taggit==1.2.0
Markdown==3.1.1
psycopg2==2.8.4 pip
sudo apt-get install libpq-dev python-dev apt
sudo apt-get install postgresql postgresql-contrib apt
